Keep wallet return scrubbing outside hosted imports
