+++ apps/web/src/hosted/billing/wallet/top-up-return.logic.ts
@@ -1,34 +1,22 @@
-import {
-	type PaymentIntentClientSecret,
-	stripeReturnPaymentIntentClientSecret,
-} from "@/hosted/billing/stripe-client-secret";
 import { SETTINGS_QUERY_KEY } from "@/lib/settings-routes";
+import { cleanWalletTopupReturnUrl, WALLET_TOPUP_RETURN_PARAM } from "@/lib/wallet-topup-return";
 
-export const WALLET_TOPUP_RETURN_PARAM = "topup_return";
-export const STRIPE_PAYMENT_INTENT_PARAM = "payment_intent";
-export const STRIPE_PAYMENT_INTENT_CLIENT_SECRET_PARAM = "payment_intent_client_secret";
-export const STRIPE_REDIRECT_STATUS_PARAM = "redirect_status";
+export {
+	bootstrapWalletTopupReturn,
+	cleanMarkedWalletTopupReturnRequest,
+	cleanWalletTopupReturnUrl,
+	consumeWalletTopupReturn,
+	coordinateWalletTopupReturn,
+	readWalletTopupReturn,
+} from "@/lib/wallet-topup-return";
 
 export type WalletTopupReturnToastKind = "info" | "error";
-
-export interface WalletTopupReturnState {
-	clientSecret: PaymentIntentClientSecret;
-}
 
 export interface WalletTopupReturnToast {
 	kind: WalletTopupReturnToastKind;
 	title: string;
 	description: string;
 }
-
-export interface WalletTopupReturnResolution {
-	status: string | null;
-	paymentIntentId: string | null;
-	errorMessage: string | null;
-}
-
-let pendingWalletTopupReturn: WalletTopupReturnState | null = null;
-let walletTopupReturnResolution: Promise<WalletTopupReturnResolution> | null = null;
 
 export const WALLET_TOPUP_ACCEPTED_TOAST = {
 	title: "Payment accepted",
@@ -39,62 +27,6 @@ export function buildWalletTopupReturnUrl(currentHref: string): string {
 	const url = new URL(cleanWalletTopupReturnUrl(currentHref));
 	url.searchParams.set(SETTINGS_QUERY_KEY, "billing-wallet");
 	url.searchParams.set(WALLET_TOPUP_RETURN_PARAM, "1");
-	return url.toString();
-}
-
-export function consumeWalletTopupReturn(
-	currentHref: string,
-	historyState: unknown,
-	replaceState: (state: unknown, unused: string, url: string) => void,
-): WalletTopupReturnState | null {
-	const url = new URL(currentHref);
-	if (!url.searchParams.has(WALLET_TOPUP_RETURN_PARAM)) return null;
-	const result = readWalletTopupReturn(url.search);
-	replaceState(historyState, "", cleanWalletTopupReturnUrl(currentHref));
-	return result;
-}
-
-export function bootstrapWalletTopupReturn(
-	currentHref: string,
-	historyState: unknown,
-	replaceState: (state: unknown, unused: string, url: string) => void,
-): void {
-	pendingWalletTopupReturn = consumeWalletTopupReturn(currentHref, historyState, replaceState);
-}
-
-export function coordinateWalletTopupReturn(
-	retrieve: (clientSecret: PaymentIntentClientSecret) => Promise<WalletTopupReturnResolution>,
-): Promise<WalletTopupReturnResolution> | null {
-	if (walletTopupReturnResolution) return walletTopupReturnResolution;
-	if (!pendingWalletTopupReturn) return null;
-	const { clientSecret } = pendingWalletTopupReturn;
-	pendingWalletTopupReturn = null;
-	walletTopupReturnResolution = retrieve(clientSecret);
-	return walletTopupReturnResolution;
-}
-
-export function cleanMarkedWalletTopupReturnRequest(request: Request): Request {
-	const url = new URL(request.url);
-	if (!url.searchParams.has(WALLET_TOPUP_RETURN_PARAM)) return request;
-	return new Request(cleanWalletTopupReturnUrl(request.url), request);
-}
-
-export function readWalletTopupReturn(search: string): WalletTopupReturnState | null {
-	const params = new URLSearchParams(search);
-	if (params.get(WALLET_TOPUP_RETURN_PARAM) !== "1") return null;
-	const clientSecret = stripeReturnPaymentIntentClientSecret(
-		params.get(STRIPE_PAYMENT_INTENT_CLIENT_SECRET_PARAM),
-	);
-	if (!clientSecret) return null;
-	return { clientSecret };
-}
-
-export function cleanWalletTopupReturnUrl(currentHref: string): string {
-	const url = new URL(currentHref);
-	url.searchParams.delete(WALLET_TOPUP_RETURN_PARAM);
-	url.searchParams.delete(STRIPE_PAYMENT_INTENT_PARAM);
-	url.searchParams.delete(STRIPE_PAYMENT_INTENT_CLIENT_SECRET_PARAM);
-	url.searchParams.delete(STRIPE_REDIRECT_STATUS_PARAM);
 	return url.toString();
 }
